Move testing reference equality snippets
