+++ docs/csharp/programming-guide/statements-expressions-operators/snippets/how-to-test-for-reference-equality-identity/TestingReferenceEquality.csproj
@@ -1,0 +1,9 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net5.0</TargetFramework>
+    <RootNamespace>TestingReferenceEquality</RootNamespace>
+  </PropertyGroup>
+
+</Project>
